Moved find_available_port to test_utils module.
